refactor(tls-client): consolidate duplicated test harness helpers
SonarCloud still reports 4.7% duplication on new code after the
HMAC/HKDF template dedup — the remaining mass is repeated test
harness code. One shared handshake/test_util.rs now owns the
ServerHello/EncryptedExtensions/Certificate/CertificateVerify
builders, the handshake-header wrapper, and the hex formatter
(previously copied across the server_hello, encrypted_extensions,
certificate_msg, key_schedule, and driver test modules); the
security crate's three hex() copies move to a shared test_util
too. Test semantics unchanged.
