+++ security/src/lib.rs
@@ -49,4 +49,6 @@ pub mod policy;
 pub mod sha1;
 pub mod sha2;
 pub mod supply_chain;
+#[cfg(test)]
+pub(crate) mod test_util;
 pub mod totp;
